seperating odh exclusive image volume code from upstream code

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
